requirements.txt
boto3
django-storages
py-staticmaps==0.2.2
mkdocs-material
